Improvement - Split models.py into multiple files.
The file `models.py` was getting to big, split it into multiple files.
