+++ gspread/worksheet.py
@@ -1,71 +1,63 @@
 # -*- coding: utf-8 -*-
 
 """
-gspread.models
+gspread.worksheet
 ~~~~~~~~~~~~~~
 
-This module contains common spreadsheets' models.
+This module contains common worksheets models.
 
 """
 
-from .exceptions import WorksheetNotFound, CellNotFound
+from .cell import (
+    Cell,
+)
 
-from .utils import (
-    a1_to_rowcol,
-    rowcol_to_a1,
-    cast_to_a1_notation,
-    numericise,
-    numericise_all,
-    finditem,
-    fill_gaps,
-    cell_list_to_rect,
-    quote,
-    is_scalar,
-    filter_dict_values,
-    absolute_range_name,
-    a1_range_to_grid_range,
-    accepted_kwargs,
+from .exceptions import (
+    CellNotFound,
 )
 
 from .urls import (
     SPREADSHEET_URL,
-    SPREADSHEET_VALUES_URL,
-    SPREADSHEET_VALUES_BATCH_URL,
-    SPREADSHEET_BATCH_UPDATE_URL,
-    SPREADSHEET_VALUES_APPEND_URL,
-    SPREADSHEET_VALUES_CLEAR_URL,
-    SPREADSHEET_DRIVE_URL,
     WORKSHEET_DRIVE_URL,
-    SPREADSHEET_VALUES_BATCH_UPDATE_URL,
-    SPREADSHEET_SHEETS_COPY_TO_URL,
 )
 from gspread import utils
 
-try:
-    unicode
-except NameError:
-    basestring = unicode = str
+from .utils import (
+    a1_range_to_grid_range,
+    a1_to_rowcol,
+    absolute_range_name,
+    accepted_kwargs,
+    basestring,
+    cast_to_a1_notation,
+    cell_list_to_rect,
+    fill_gaps,
+    filter_dict_values,
+    finditem,
+    is_scalar,
+    numericise_all,
+    rowcol_to_a1,
+)
 
 
 class ValueRange(list):
     @classmethod
     def from_json(cls, json):
-        values = json.get('values', [])
+        values = json.get("values", [])
         new_obj = cls(values)
         new_obj._json = {
-            'range': json['range'],
-            'majorDimension': json['majorDimension'],
+            "range": json["range"],
+            "majorDimension": json["majorDimension"],
         }
 
         return new_obj
 
     @property
     def range(self):
-        return self._json['range']
+        return self._json["range"]
 
     @property
     def major_dimension(self):
-        return self._json['majorDimension']
+        return self._json["majorDimension"]
 
     def first(self, default=None):
         """Returns the value of a first cell in a range.
@@ -76,494 +68,6 @@ class ValueRange(list):
             return self[0][0]
         except IndexError:
             return default
-
-
-class Spreadsheet(object):
-    """The class that represents a spreadsheet."""
-
-    def __init__(self, client, properties):
-        self.client = client
-        self._properties = properties
-
-    @property
-    def id(self):
-        """Spreadsheet ID."""
-        return self._properties['id']
-
-    @property
-    def title(self):
-        """Spreadsheet title."""
-        try:
-            return self._properties['title']
-        except KeyError:
-            metadata = self.fetch_sheet_metadata()
-            self._properties.update(metadata['properties'])
-            return self._properties['title']
-
-    @property
-    def url(self):
-        """Spreadsheet URL."""
-        return SPREADSHEET_DRIVE_URL % self.id
-
-    @property
-    def creationTime(self):
-        """Spreadsheet Creation time."""
-        try:
-            return self._properties['createdTime']
-        except KeyError:
-            # Filter the list using the name to reduce the request size
-            # Filter the item using the unique ID to ensure we update the exacte same item
-            metadata = utils.finditem(lambda x: x["id"] == self.id,
-                                    self.client.list_spreadsheet_files(self.title))
-            self._properties.update(metadata)
-            return self._properties['createdTime']
-
-    @property
-    def lastUpdateTime(self):
-        """Spreadsheet Creation time."""
-        try:
-            return self._properties['modifiedTime']
-        except KeyError:
-            # Filter the list using the name to reduce the request size
-            # Filter the item using the unique ID to ensure we update the exacte same item
-            metadata = utils.finditem(lambda x: x["id"] == self.id,
-                                    self.client.list_spreadsheet_files(self.title))
-            self._properties.update(metadata)
-            return self._properties['modifiedTime']
-
-    @property
-    def updated(self):
-        """.. deprecated:: 2.0
-
-        This feature is not supported in Sheets API v4.
-        """
-        import warnings
-
-        warnings.warn(
-            "Spreadsheet.updated() is deprecated, "
-            "this feature is not supported in Sheets API v4",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-    @property
-    def sheet1(self):
-        """Shortcut property for getting the first worksheet."""
-        return self.get_worksheet(0)
-
-    def __iter__(self):
-        for sheet in self.worksheets():
-            yield (sheet)
-
-    def __repr__(self):
-        return '<%s %s id:%s>' % (
-            self.__class__.__name__,
-            repr(self.title),
-            self.id,
-        )
-
-    def batch_update(self, body):
-        """Lower-level method that directly calls `spreadsheets.batchUpdate <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/batchUpdate>`_.
-
-        :param dict body: `Request body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/batchUpdate#request-body>`_.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/batchUpdate#response-body>`_.
-        :rtype: dict
-
-        .. versionadded:: 3.0
-        """
-        r = self.client.request(
-            'post', SPREADSHEET_BATCH_UPDATE_URL % self.id, json=body
-        )
-
-        return r.json()
-
-    def values_append(self, range, params, body):
-        """Lower-level method that directly calls `spreadsheets.values.append <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/append>`_.
-
-        :param str range: The `A1 notation <https://developers.google.com/sheets/api/guides/concepts#a1_notation>`_
-                          of a range to search for a logical table of data. Values will be appended after the last row of the table.
-        :param dict params: `Query parameters <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/append#query-parameters>`_.
-        :param dict body: `Request body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/append#request-body>`_.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/append#response-body>`_.
-        :rtype: dict
-
-        .. versionadded:: 3.0
-        """
-        url = SPREADSHEET_VALUES_APPEND_URL % (self.id, quote(range))
-        r = self.client.request('post', url, params=params, json=body)
-        return r.json()
-
-    def values_clear(self, range):
-        """Lower-level method that directly calls `spreadsheets.values.clear <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/clear>`_.
-
-        :param str range: The `A1 notation <https://developers.google.com/sheets/api/guides/concepts#a1_notation>`_ of the values to clear.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/clear#response-body>`_.
-        :rtype: dict
-
-        .. versionadded:: 3.0
-        """
-        url = SPREADSHEET_VALUES_CLEAR_URL % (self.id, quote(range))
-        r = self.client.request('post', url)
-        return r.json()
-
-    def values_get(self, range, params=None):
-        """Lower-level method that directly calls `spreadsheets.values.get <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get>`_.
-
-        :param str range: The `A1 notation <https://developers.google.com/sheets/api/guides/concepts#a1_notation>`_ of the values to retrieve.
-        :param dict params: (optional) `Query parameters <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get#query-parameters>`_.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get#response-body>`_.
-        :rtype: dict
-
-        .. versionadded:: 3.0
-        """
-        url = SPREADSHEET_VALUES_URL % (self.id, quote(range))
-        r = self.client.request('get', url, params=params)
-        return r.json()
-
-    def values_batch_get(self, ranges, params=None):
-        """Lower-level method that directly calls `spreadsheets.values.batchGet <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/batchGet>`_.
-
-        :param ranges: List of ranges in the `A1 notation <https://developers.google.com/sheets/api/guides/concepts#a1_notation>`_ of the values to retrieve.
-        :param dict params: (optional) `Query parameters <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get#query-parameters>`_.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get#response-body>`_.
-        :rtype: dict
-        """
-        if params is None:
-            params = {}
-
-        params.update(ranges=ranges)
-
-        url = SPREADSHEET_VALUES_BATCH_URL % (self.id)
-        r = self.client.request("get", url, params=params)
-        return r.json()
-
-    def values_update(self, range, params=None, body=None):
-        """Lower-level method that directly calls `spreadsheets.values.update <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/update>`_.
-
-        :param str range: The `A1 notation <https://developers.google.com/sheets/api/guides/concepts#a1_notation>`_ of the values to update.
-        :param dict params: (optional) `Query parameters <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/update#query-parameters>`_.
-        :param dict body: (optional) `Request body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/update#request-body>`_.
-        :returns: `Response body <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/update#response-body>`_.
-        :rtype: dict
-
-        Example::
-
-            sh.values_update(
-                'Sheet1!A2',
-                params={
-                    'valueInputOption': 'USER_ENTERED'
-                },
-                body={
-                    'values': [[1, 2, 3]]
-                }
-            )
-
-        .. versionadded:: 3.0
-        """
-        url = SPREADSHEET_VALUES_URL % (self.id, quote(range))
-        r = self.client.request('put', url, params=params, json=body)
-        return r.json()
-
-    def values_batch_update(self, params=None, body=None):
-        url = SPREADSHEET_VALUES_BATCH_UPDATE_URL % self.id
-        r = self.client.request('post', url, params=params, json=body)
-        return r.json()
-
-    def _spreadsheets_get(self, params=None):
-        """A method stub that directly calls `spreadsheets.get <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/get>`_. """
-        url = SPREADSHEET_URL % self.id
-        r = self.client.request('get', url, params=params)
-        return r.json()
-
-    def _spreadsheets_sheets_copy_to(
-        self, sheet_id, destination_spreadsheet_id
-    ):
-        """Lower-level method that directly calls `spreadsheets.sheets.copyTo <https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.sheets/copyTo>`_. """
-        url = SPREADSHEET_SHEETS_COPY_TO_URL % (self.id, sheet_id)
-
-        body = {"destinationSpreadsheetId": destination_spreadsheet_id}
-        r = self.client.request('post', url, json=body)
-        return r.json()
-
-    def fetch_sheet_metadata(self, params=None):
-        if params is None:
-            params = {'includeGridData': 'false'}
-
-        url = SPREADSHEET_URL % self.id
-
-        r = self.client.request('get', url, params=params)
-
-        return r.json()
-
-    def get_worksheet(self, index):
-        """Returns a worksheet with specified `index`.
-
-        :param index: An index of a worksheet. Indexes start from zero.
-        :type index: int
-
-        :returns: an instance of :class:`gspread.models.Worksheet`
-                  or `None` if the worksheet is not found.
-
-        Example. To get first worksheet of a spreadsheet:
-
-        >>> sht = client.open('My fancy spreadsheet')
-        >>> worksheet = sht.get_worksheet(0):rtype: dict
-        """
-        sheet_data = self.fetch_sheet_metadata()
-
-        try:
-            properties = sheet_data['sheets'][index]['properties']
-            return Worksheet(self, properties)
-        except (KeyError, IndexError):
-            return None
-
-    def worksheets(self):
-        """Returns a list of all :class:`worksheets <gspread.models.Worksheet>`
-        in a spreadsheet.
-        """
-        sheet_data = self.fetch_sheet_metadata()
-        return [Worksheet(self, x['properties']) for x in sheet_data['sheets']]
-
-    def worksheet(self, title):
-        """Returns a worksheet with specified `title`.
-
-        :param title: A title of a worksheet. If there're multiple
-                      worksheets with the same title, first one will
-                      be returned.
-        :type title: str
-
-        :returns: an instance of :class:`gspread.models.Worksheet`.
-
-        Example. Getting worksheet named 'Annual bonuses'
-
-        >>> sht = client.open('Sample one')
-        >>> worksheet = sht.worksheet('Annual bonuses')
-        """
-        sheet_data = self.fetch_sheet_metadata()
-        try:
-            item = finditem(
-                lambda x: x['properties']['title'] == title,
-                sheet_data['sheets'],
-            )
-            return Worksheet(self, item['properties'])
-        except (StopIteration, KeyError):
-            raise WorksheetNotFound(title)
-
-    def add_worksheet(self, title, rows, cols, index=None):
-        """Adds a new worksheet to a spreadsheet.
-
-        :param title: A title of a new worksheet.
-        :type title: str
-        :param rows: Number of rows.
-        :type rows: int
-        :param cols: Number of columns.
-        :type cols: int
-        :param index: Position of the sheet.
-        :type index: int
-
-        :returns: a newly created :class:`worksheets <gspread.models.Worksheet>`.
-        """
-        body = {
-            'requests': [
-                {
-                    'addSheet': {
-                        'properties': {
-                            'title': title,
-                            'sheetType': 'GRID',
-                            'gridProperties': {
-                                'rowCount': rows,
-                                'columnCount': cols,
-                            },
-                        }
-                    }
-                }
-            ]
-        }
-
-        if index is not None:
-            body["requests"][0]["addSheet"]["properties"]["index"] = index
-
-        data = self.batch_update(body)
-
-        properties = data['replies'][0]['addSheet']['properties']
-
-        worksheet = Worksheet(self, properties)
-
-        return worksheet
-
-    def duplicate_sheet(
-        self,
-        source_sheet_id,
-        insert_sheet_index=None,
-        new_sheet_id=None,
-        new_sheet_name=None,
-    ):
-        """Duplicates the contents of a sheet.
-
-        :param int source_sheet_id: The sheet ID to duplicate.
-        :param int insert_sheet_index: (optional) The zero-based index
-                                       where the new sheet should be inserted.
-                                       The index of all sheets after this are
-                                       incremented.
-        :param int new_sheet_id: (optional) The ID of the new sheet.
-                                 If not set, an ID is chosen. If set, the ID
-                                 must not conflict with any existing sheet ID.
-                                 If set, it must be non-negative.
-        :param str new_sheet_name: (optional) The name of the new sheet.
-                                   If empty, a new name is chosen for you.
-
-        :returns: a newly created :class:`<gspread.models.Worksheet>`.
-
-        .. versionadded:: 3.1
-        """
-        body = {
-            'requests': [
-                {
-                    'duplicateSheet': {
-                        'sourceSheetId': source_sheet_id,
-                        'insertSheetIndex': insert_sheet_index,
-                        'newSheetId': new_sheet_id,
-                        'newSheetName': new_sheet_name,
-                    }
-                }
-            ]
-        }
-
-        data = self.batch_update(body)
-
-        properties = data['replies'][0]['duplicateSheet']['properties']
-
-        worksheet = Worksheet(self, properties)
-
-        return worksheet
-
-    def del_worksheet(self, worksheet):
-        """Deletes a worksheet from a spreadsheet.
-
-        :param worksheet: The worksheet to be deleted.
-        :type worksheet: :class:`~gspread.Worksheet`
-        """
-        body = {'requests': [{'deleteSheet': {'sheetId': worksheet.id}}]}
-
-        return self.batch_update(body)
-
-    def reorder_worksheets(self, worksheets_in_desired_order):
-        """Updates the ``index`` property of each Worksheets to reflect
-        its index in the provided sequence of Worksheets.
-
-        :param worksheets_in_desired_order: Iterable of Worksheet objects in desired order.
-
-        Note: If you omit some of the Spreadsheet's existing Worksheet objects from
-        the provided sequence, those Worksheets will be appended to the end of the sequence
-        in the order that they appear in the list returned by ``Spreadsheet.worksheets()``.
-
-        .. versionadded:: 3.4
-        """
-        idx_map = {}
-        for idx, w in enumerate(worksheets_in_desired_order):
-            idx_map[w.id] = idx
-        for w in self.worksheets():
-            if w.id in idx_map:
-                continue
-            idx += 1
-            idx_map[w.id] = idx
-
-        body = {
-            'requests': [
-                {
-                    'updateSheetProperties': {
-                        'properties': {'sheetId': key, 'index': val},
-                        'fields': 'index',
-                    }
-                }
-                for key, val in idx_map.items()
-            ]
-        }
-
-        return self.batch_update(body)
-
-    def share(
-        self,
-        value,
-        perm_type,
-        role,
-        notify=True,
-        email_message=None,
-        with_link=False,
-    ):
-        """Share the spreadsheet with other accounts.
-
-        :param value: user or group e-mail address, domain name
-                      or None for 'default' type.
-        :type value: str, None
-        :param perm_type: The account type.
-               Allowed values are: ``user``, ``group``, ``domain``,
-               ``anyone``.
-        :type perm_type: str
-        :param role: The primary role for this user.
-               Allowed values are: ``owner``, ``writer``, ``reader``.
-        :type role: str
-        :param notify: (optional) Whether to send an email to the target user/domain.
-        :type notify: str
-        :param email_message: (optional) The email to be sent if notify=True
-        :type email_message: str
-
-        :param with_link: (optional) Whether the link is required for this permission
-        :type with_link: bool
-
-        Example::
-
-            # Give Otto a write permission on this spreadsheet
-            sh.share('otto@example.com', perm_type='user', role='writer')
-
-            # Transfer ownership to Otto
-            sh.share('otto@example.com', perm_type='user', role='owner')
-        """
-        self.client.insert_permission(
-            self.id,
-            value=value,
-            perm_type=perm_type,
-            role=role,
-            notify=notify,
-            email_message=email_message,
-            with_link=with_link,
-        )
-
-    def list_permissions(self):
-        """Lists the spreadsheet's permissions."""
-        return self.client.list_permissions(self.id)
-
-    def remove_permissions(self, value, role='any'):
-        """Remove permissions from a user or domain.
-
-        :param value: User or domain to remove permissions from
-        :type value: str
-        :param role: (optional) Permission to remove. Defaults to all
-                     permissions.
-        :type role: str
-
-        Example::
-
-            # Remove Otto's write permission for this spreadsheet
-            sh.remove_permissions('otto@example.com', role='writer')
-
-            # Remove all Otto's permissions for this spreadsheet
-            sh.remove_permissions('otto@example.com')
-        """
-        permission_list = self.client.list_permissions(self.id)
-
-        key = 'emailAddress' if '@' in value else 'domain'
-
-        filtered_id_list = [
-            p['id']
-            for p in permission_list
-            if p.get(key) == value and (p['role'] == role or role == 'any')
-        ]
-
-        for permission_id in filtered_id_list:
-            self.client.remove_permission(self.id, permission_id)
-
-        return filtered_id_list
 
 
 class Worksheet(object):
@@ -577,7 +81,7 @@ class Worksheet(object):
         self._properties = properties
 
     def __repr__(self):
-        return '<%s %s id:%s>' % (
+        return "<%s %s id:%s>" % (
             self.__class__.__name__,
             repr(self.title),
             self.id,
@@ -586,12 +90,12 @@ class Worksheet(object):
     @property
     def id(self):
         """Worksheet ID."""
-        return self._properties['sheetId']
+        return self._properties["sheetId"]
 
     @property
     def title(self):
         """Worksheet title."""
-        return self._properties['title']
+        return self._properties["title"]
 
     @property
     def url(self):
@@ -616,24 +120,24 @@ class Worksheet(object):
     @property
     def row_count(self):
         """Number of rows."""
-        return self._properties['gridProperties']['rowCount']
+        return self._properties["gridProperties"]["rowCount"]
 
     @property
     def col_count(self):
         """Number of columns."""
-        return self._properties['gridProperties']['columnCount']
+        return self._properties["gridProperties"]["columnCount"]
 
     @property
     def frozen_row_count(self):
         """Number of frozen rows."""
-        return self._properties['gridProperties'].get('frozenRowCount', 0)
+        return self._properties["gridProperties"].get("frozenRowCount", 0)
 
     @property
     def frozen_col_count(self):
         """Number of frozen columns."""
-        return self._properties['gridProperties'].get('frozenColumnCount', 0)
+        return self._properties["gridProperties"].get("frozenColumnCount", 0)
 
-    def acell(self, label, value_render_option='FORMATTED_VALUE'):
+    def acell(self, label, value_render_option="FORMATTED_VALUE"):
         """Returns an instance of a :class:`gspread.models.Cell`.
 
         :param label: Cell label in A1 notation
@@ -655,7 +159,7 @@ class Worksheet(object):
             *(a1_to_rowcol(label)), value_render_option=value_render_option
         )
 
-    def cell(self, row, col, value_render_option='FORMATTED_VALUE'):
+    def cell(self, row, col, value_render_option="FORMATTED_VALUE"):
         """Returns an instance of a :class:`gspread.models.Cell` located at
         `row` and `col` column.
 
@@ -682,7 +186,7 @@ class Worksheet(object):
 
             value = data.first()
         except KeyError:
-            value = ''
+            value = ""
 
         return Cell(row, col, value)
 
@@ -715,11 +219,11 @@ class Worksheet(object):
 
         data = self.spreadsheet.values_get(range_label)
 
-        start, end = name.split(':')
+        start, end = name.split(":")
         (row_offset, column_offset) = a1_to_rowcol(start)
         (last_row, last_column) = a1_to_rowcol(end)
 
-        values = data.get('values', [])
+        values = data.get("values", [])
 
         rect_values = fill_gaps(
             values,
@@ -877,7 +381,7 @@ class Worksheet(object):
 
         keys = data[idx]
 
-        if numericise_ignore == ['all']:
+        if numericise_ignore == ["all"]:
             values = data[idx + 1 :]
         else:
             values = [
@@ -911,12 +415,12 @@ class Worksheet(object):
         .. _ValueRenderOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueRenderOption
         """
         try:
-            data = self.get('A{}:{}'.format(row, row), **kwargs)
+            data = self.get("A{}:{}".format(row, row), **kwargs)
             return data[0]
         except KeyError:
             return []
 
-    def col_values(self, col, value_render_option='FORMATTED_VALUE'):
+    def col_values(self, col, value_render_option="FORMATTED_VALUE"):
         """Returns a list of all values in column `col`.
 
         Empty cells in this list will be rendered as :const:`None`.
@@ -930,20 +434,20 @@ class Worksheet(object):
         """
 
         start_label = rowcol_to_a1(1, col)
-        range_label = '{}:{}'.format(start_label, start_label[:-1])
+        range_label = "{}:{}".format(start_label, start_label[:-1])
 
         range_name = absolute_range_name(self.title, range_label)
 
         data = self.spreadsheet.values_get(
             range_name,
             params={
-                'valueRenderOption': value_render_option,
-                'majorDimension': 'COLUMNS',
+                "valueRenderOption": value_render_option,
+                "majorDimension": "COLUMNS",
             },
         )
 
         try:
-            return data['values'][0]
+            return data["values"][0]
         except KeyError:
             return []
 
@@ -974,13 +478,13 @@ class Worksheet(object):
 
         data = self.spreadsheet.values_update(
             range_name,
-            params={'valueInputOption': 'USER_ENTERED'},
-            body={'values': [[value]]},
+            params={"valueInputOption": "USER_ENTERED"},
+            body={"values": [[value]]},
         )
 
         return data
 
-    def update_cells(self, cell_list, value_input_option='RAW'):
+    def update_cells(self, cell_list, value_input_option="RAW"):
         """Updates many cells at once.
 
         :param list cell_list: List of :class:`Cell` objects to update.
@@ -1018,18 +522,14 @@ class Worksheet(object):
         start = rowcol_to_a1(
             min(c.row for c in cell_list), min(c.col for c in cell_list)
         )
-        end = rowcol_to_a1(
-            max(c.row for c in cell_list), max(c.col for c in cell_list)
-        )
+        end = rowcol_to_a1(max(c.row for c in cell_list), max(c.col for c in cell_list))
 
-        range_name = absolute_range_name(
-            self.title, '{}:{}'.format(start, end)
-        )
+        range_name = absolute_range_name(self.title, "{}:{}".format(start, end))
 
         data = self.spreadsheet.values_update(
             range_name,
-            params={'valueInputOption': value_input_option},
-            body={'values': values_rect},
+            params={"valueInputOption": value_input_option},
+            body={"values": values_rect},
         )
 
         return data
@@ -1077,9 +577,9 @@ class Worksheet(object):
 
         params = filter_dict_values(
             {
-                'majorDimension': kwargs['major_dimension'],
-                'valueRenderOption': kwargs['value_render_option'],
-                'dateTimeRenderOption': kwargs['date_time_render_option'],
+                "majorDimension": kwargs["major_dimension"],
+                "valueRenderOption": kwargs["value_render_option"],
+                "dateTimeRenderOption": kwargs["date_time_render_option"],
             }
         )
 
@@ -1121,17 +621,15 @@ class Worksheet(object):
 
         params = filter_dict_values(
             {
-                'majorDimension': kwargs['major_dimension'],
-                'valueRenderOption': kwargs['value_render_option'],
-                'dateTimeRenderOption': kwargs['date_time_render_option'],
+                "majorDimension": kwargs["major_dimension"],
+                "valueRenderOption": kwargs["value_render_option"],
+                "dateTimeRenderOption": kwargs["date_time_render_option"],
             }
         )
 
-        response = self.spreadsheet.values_batch_get(
-            ranges=ranges, params=params
-        )
+        response = self.spreadsheet.values_batch_get(ranges=ranges, params=params)
 
-        return [ValueRange.from_json(x) for x in response['valueRanges']]
+        return [ValueRange.from_json(x) for x in response["valueRanges"]]
 
     @accepted_kwargs(
         raw=True,
@@ -1203,22 +701,16 @@ class Worksheet(object):
         if is_scalar(values):
             values = [[values]]
 
-        if not kwargs['value_input_option']:
-            kwargs['value_input_option'] = (
-                'RAW' if kwargs['raw'] else 'USER_ENTERED'
-            )
+        if not kwargs["value_input_option"]:
+            kwargs["value_input_option"] = "RAW" if kwargs["raw"] else "USER_ENTERED"
 
         params = filter_dict_values(
             {
-                'valueInputOption': kwargs['value_input_option'],
-                'includeValuesInResponse': kwargs[
-                    'include_values_in_response'
-                ],
-                'responseValueRenderOption': kwargs[
-                    'response_value_render_option'
-                ],
-                'responseDateTimeRenderOption': kwargs[
-                    'response_date_time_render_option'
+                "valueInputOption": kwargs["value_input_option"],
+                "includeValuesInResponse": kwargs["include_values_in_response"],
+                "responseValueRenderOption": kwargs["response_value_render_option"],
+                "responseDateTimeRenderOption": kwargs[
+                    "response_date_time_render_option"
                 ],
             }
         )
@@ -1227,7 +719,7 @@ class Worksheet(object):
             range_name,
             params=params,
             body=filter_dict_values(
-                {'values': values, 'majorDimension': kwargs['major_dimension']}
+                {"values": values, "majorDimension": kwargs["major_dimension"]}
             ),
         )
 
@@ -1281,29 +773,22 @@ class Worksheet(object):
 
         .. versionadded:: 3.3
         """
-        if not kwargs['value_input_option']:
-            kwargs['value_input_option'] = (
-                'RAW' if kwargs['raw'] else 'USER_ENTERED'
-            )
+        if not kwargs["value_input_option"]:
+            kwargs["value_input_option"] = "RAW" if kwargs["raw"] else "USER_ENTERED"
 
         data = [
-            dict(vr, range=absolute_range_name(self.title, vr['range']))
-            for vr in data
+            dict(vr, range=absolute_range_name(self.title, vr["range"])) for vr in data
         ]
 
         body = filter_dict_values(
             {
-                'valueInputOption': kwargs['value_input_option'],
-                'includeValuesInResponse': kwargs[
-                    'include_values_in_response'
+                "valueInputOption": kwargs["value_input_option"],
+                "includeValuesInResponse": kwargs["include_values_in_response"],
+                "responseValueRenderOption": kwargs["response_value_render_option"],
+                "responseDateTimeRenderOption": kwargs[
+                    "response_date_time_render_option"
                 ],
-                'responseValueRenderOption': kwargs[
-                    'response_value_render_option'
-                ],
-                'responseDateTimeRenderOption': kwargs[
-                    'response_date_time_render_option'
-                ],
-                'data': data,
+                "data": data,
             }
         )
 
@@ -1349,7 +834,7 @@ class Worksheet(object):
         """
         grid_range = a1_range_to_grid_range(range_name, self.id)
 
-        fields = "userEnteredFormat(%s)" % ','.join(cell_format.keys())
+        fields = "userEnteredFormat(%s)" % ",".join(cell_format.keys())
 
         body = {
             "requests": [
@@ -1374,27 +859,25 @@ class Worksheet(object):
         grid_properties = {}
 
         if rows is not None:
-            grid_properties['rowCount'] = rows
+            grid_properties["rowCount"] = rows
 
         if cols is not None:
-            grid_properties['columnCount'] = cols
+            grid_properties["columnCount"] = cols
 
         if not grid_properties:
             raise TypeError("Either 'rows' or 'cols' should be specified.")
 
-        fields = ','.join(
-            'gridProperties/%s' % p for p in grid_properties.keys()
-        )
+        fields = ",".join("gridProperties/%s" % p for p in grid_properties.keys())
 
         body = {
-            'requests': [
+            "requests": [
                 {
-                    'updateSheetProperties': {
-                        'properties': {
-                            'sheetId': self.id,
-                            'gridProperties': grid_properties,
+                    "updateSheetProperties": {
+                        "properties": {
+                            "sheetId": self.id,
+                            "gridProperties": grid_properties,
                         },
-                        'fields': fields,
+                        "fields": fields,
                     }
                 }
             ]
@@ -1424,34 +907,32 @@ class Worksheet(object):
 
         .. versionadded:: 3.4
         """
-        range_name = kwargs.pop('range', None)
+        range_name = kwargs.pop("range", None)
 
         if range_name:
-            start_a1, end_a1 = range_name.split(':')
+            start_a1, end_a1 = range_name.split(":")
             start_row, start_col = a1_to_rowcol(start_a1)
             end_row, end_col = a1_to_rowcol(end_a1)
         else:
-            start_row = (
-                self._properties['gridProperties'].get('frozenRowCount', 0) + 1
-            )
+            start_row = self._properties["gridProperties"].get("frozenRowCount", 0) + 1
             start_col = 1
             end_row = self.row_count
             end_col = self.col_count
 
         request_range = {
-            'sheetId': self.id,
-            'startRowIndex': start_row - 1,
-            'endRowIndex': end_row,
-            'startColumnIndex': start_col - 1,
-            'endColumnIndex': end_col,
+            "sheetId": self.id,
+            "startRowIndex": start_row - 1,
+            "endRowIndex": end_row,
+            "startColumnIndex": start_col - 1,
+            "endColumnIndex": end_col,
         }
 
         request_sort_specs = list()
         for col, order in specs:
-            if order == 'asc':
-                request_order = 'ASCENDING'
-            elif order == 'des':
-                request_order = 'DESCENDING'
+            if order == "asc":
+                request_order = "ASCENDING"
+            elif order == "des":
+                request_order = "DESCENDING"
             else:
                 raise ValueError(
                     "Either 'asc' or 'des' should be specified as sort order."
@@ -1463,11 +944,11 @@ class Worksheet(object):
             request_sort_specs.append(request_sort_spec)
 
         body = {
-            'requests': [
+            "requests": [
                 {
                     "sortRange": {
-                        'range': request_range,
-                        'sortSpecs': request_sort_specs,
+                        "range": request_range,
+                        "sortSpecs": request_sort_specs,
                     }
                 }
             ]
@@ -1482,18 +963,18 @@ class Worksheet(object):
         :param str title: A new title.
         """
         body = {
-            'requests': [
+            "requests": [
                 {
-                    'updateSheetProperties': {
-                        'properties': {'sheetId': self.id, 'title': title},
-                        'fields': 'title',
+                    "updateSheetProperties": {
+                        "properties": {"sheetId": self.id, "title": title},
+                        "fields": "title",
                     }
                 }
             ]
         }
 
         response = self.spreadsheet.batch_update(body)
-        self._properties['title'] = title
+        self._properties["title"] = title
         return response
 
     def update_index(self, index):
@@ -1509,11 +990,11 @@ class Worksheet(object):
         .. versionadded:: 3.4
         """
         body = {
-            'requests': [
+            "requests": [
                 {
-                    'updateSheetProperties': {
-                        'properties': {'sheetId': self.id, 'index': index},
-                        'fields': 'index',
+                    "updateSheetProperties": {
+                        "properties": {"sheetId": self.id, "index": index},
+                        "fields": "index",
                     }
                 }
             ]
@@ -1542,7 +1023,7 @@ class Worksheet(object):
     def append_row(
         self,
         values,
-        value_input_option='RAW',
+        value_input_option="RAW",
         insert_data_option=None,
         table_range=None,
     ):
@@ -1575,7 +1056,7 @@ class Worksheet(object):
     def append_rows(
         self,
         values,
-        value_input_option='RAW',
+        value_input_option="RAW",
         insert_data_option=None,
         table_range=None,
     ):
@@ -1601,15 +1082,15 @@ class Worksheet(object):
         range_label = absolute_range_name(self.title, table_range)
 
         params = {
-            'valueInputOption': value_input_option,
-            'insertDataOption': insert_data_option,
+            "valueInputOption": value_input_option,
+            "insertDataOption": insert_data_option,
         }
 
-        body = {'values': values}
+        body = {"values": values}
 
         return self.spreadsheet.values_append(range_label, params, body)
 
-    def insert_row(self, values, index=1, value_input_option='RAW'):
+    def insert_row(self, values, index=1, value_input_option="RAW"):
         """Adds a row to the worksheet at the specified index and populates it
         with values.
 
@@ -1625,7 +1106,7 @@ class Worksheet(object):
         """
         return self.insert_rows([values], index, value_input_option=value_input_option)
 
-    def insert_rows(self, values, row=1, value_input_option='RAW'):
+    def insert_rows(self, values, row=1, value_input_option="RAW"):
         """Adds multiple rows to the worksheet at the specified index and
         populates them with values.
 
@@ -1654,15 +1135,15 @@ class Worksheet(object):
 
         self.spreadsheet.batch_update(body)
 
-        range_label = absolute_range_name(self.title, 'A%s' % row)
+        range_label = absolute_range_name(self.title, "A%s" % row)
 
-        params = {'valueInputOption': value_input_option}
+        params = {"valueInputOption": value_input_option}
 
-        body = {'majorDimension': 'ROWS', 'values': values}
+        body = {"majorDimension": "ROWS", "values": values}
 
         return self.spreadsheet.values_append(range_label, params, body)
 
-    def insert_cols(self, values, col=1, value_input_option='RAW'):
+    def insert_cols(self, values, col=1, value_input_option="RAW"):
         """Adds multiple new cols to the worksheet at specified index and
         populates them with values.
 
@@ -1693,9 +1174,9 @@ class Worksheet(object):
 
         range_label = absolute_range_name(self.title, rowcol_to_a1(1, col))
 
-        params = {'valueInputOption': value_input_option}
+        params = {"valueInputOption": value_input_option}
 
-        body = {'majorDimension': 'COLUMNS', 'values': values}
+        body = {"majorDimension": "COLUMNS", "values": values}
 
         return self.spreadsheet.values_append(range_label, params, body)
 
@@ -1724,7 +1205,7 @@ class Worksheet(object):
         warning_only=False,
         requesting_user_can_edit=False,
     ):
-        """"Add protected range to the sheet. Only the editors can edit
+        """ "Add protected range to the sheet. Only the editors can edit
         the protected range.
 
         :param str name: A string with range value in A1 notation,
@@ -1753,9 +1234,9 @@ class Worksheet(object):
             Defaults to ``False``.
         """
         email_address = [
-            permission.get('emailAddress')
+            permission.get("emailAddress")
             for permission in self.client.list_permissions(self.spreadsheet.id)
-            if permission.get('emailAddress')
+            if permission.get("emailAddress")
         ]
 
         editors_emails = editor_users_emails or []
@@ -1769,7 +1250,7 @@ class Worksheet(object):
             "requests": [
                 {
                     "addProtectedRange": {
-                        'protectedRange': {
+                        "protectedRange": {
                             "range": grid_range,
                             "description": description,
                             "warningOnly": warning_only,
@@ -1832,7 +1313,7 @@ class Worksheet(object):
             worksheet.delete_rows(2)
 
         """
-        return self.delete_dimension('ROWS', start_index, end_index)
+        return self.delete_dimension("ROWS", start_index, end_index)
 
     def delete_columns(self, start_index, end_index=None):
         """Deletes multiple columns from the worksheet at the specified index.
@@ -1842,7 +1323,7 @@ class Worksheet(object):
             When end_index is not specified this method only deletes a single
             column at ``start_index``.
         """
-        return self.delete_dimension('COLUMNS', start_index, end_index)
+        return self.delete_dimension("COLUMNS", start_index, end_index)
 
     def clear(self):
         """Clears all cells in the worksheet."""
@@ -1852,7 +1333,7 @@ class Worksheet(object):
         data = self.spreadsheet.values_get(absolute_range_name(self.title))
 
         try:
-            values = fill_gaps(data['values'])
+            values = fill_gaps(data["values"])
         except KeyError:
             values = []
 
@@ -1870,9 +1351,7 @@ class Worksheet(object):
         ``in_row``` or ``in_column`` values (both one-based).
         """
         if in_row and in_column:
-            raise TypeError(
-                "Either 'in_row' or 'in_column' should be specified."
-            )
+            raise TypeError("Either 'in_row' or 'in_column' should be specified.")
 
         if in_column:
             return [
@@ -1925,27 +1404,25 @@ class Worksheet(object):
         grid_properties = {}
 
         if rows is not None:
-            grid_properties['frozenRowCount'] = rows
+            grid_properties["frozenRowCount"] = rows
 
         if cols is not None:
-            grid_properties['frozenColumnCount'] = cols
+            grid_properties["frozenColumnCount"] = cols
 
         if not grid_properties:
             raise TypeError("Either 'rows' or 'cols' should be specified.")
 
-        fields = ','.join(
-            'gridProperties/%s' % p for p in grid_properties.keys()
-        )
+        fields = ",".join("gridProperties/%s" % p for p in grid_properties.keys())
 
         body = {
-            'requests': [
+            "requests": [
                 {
-                    'updateSheetProperties': {
-                        'properties': {
-                            'sheetId': self.id,
-                            'gridProperties': grid_properties,
+                    "updateSheetProperties": {
+                        "properties": {
+                            "sheetId": self.id,
+                            "gridProperties": grid_properties,
                         },
-                        'fields': fields,
+                        "fields": fields,
                     }
                 }
             ]
@@ -1974,12 +1451,10 @@ class Worksheet(object):
         grid_range = (
             a1_range_to_grid_range(name, self.id)
             if name is not None
-            else {'sheetId': self.id}
+            else {"sheetId": self.id}
         )
 
-        body = {
-            'requests': [{'setBasicFilter': {'filter': {'range': grid_range}}}]
-        }
+        body = {"requests": [{"setBasicFilter": {"filter": {"range": grid_range}}}]}
 
         return self.spreadsheet.batch_update(body)
 
@@ -1988,7 +1463,15 @@ class Worksheet(object):
 
         .. versionadded:: 3.4
         """
-        body = {'requests': [{'clearBasicFilter': {'sheetId': self.id,}}]}
+        body = {
+            "requests": [
+                {
+                    "clearBasicFilter": {
+                        "sheetId": self.id,
+                    }
+                }
+            ]
+        }
 
         return self.spreadsheet.batch_update(body)
 
@@ -2029,7 +1512,8 @@ class Worksheet(object):
         )
 
     def copy_to(
-        self, spreadsheet_id,
+        self,
+        spreadsheet_id,
     ):
         """Copies this sheet to another spreadsheet.
 
@@ -2039,9 +1523,7 @@ class Worksheet(object):
             the newly created sheet.
         :rtype: dict
         """
-        return self.spreadsheet._spreadsheets_sheets_copy_to(
-            self.id, spreadsheet_id
-        )
+        return self.spreadsheet._spreadsheets_sheets_copy_to(self.id, spreadsheet_id)
 
     @cast_to_a1_notation
     def merge_cells(self, name, merge_type="MERGE_ALL"):
@@ -2070,9 +1552,7 @@ class Worksheet(object):
         grid_range = a1_range_to_grid_range(name, self.id)
 
         body = {
-            "requests": [
-                {"mergeCells": {"mergeType": merge_type, "range": grid_range}}
-            ]
+            "requests": [{"mergeCells": {"mergeType": merge_type, "range": grid_range}}]
         }
 
         return self.spreadsheet.batch_update(body)
@@ -2086,18 +1566,17 @@ class Worksheet(object):
         """
         absolute_cell = absolute_range_name(self.title, cell)
         url = SPREADSHEET_URL % (self.spreadsheet.id)
-        params = {
-            'ranges': absolute_cell,
-            'fields': 'sheets/data/rowData/values/note'
-        }
-        response = self.client.request('get', url, params=params)
+        params = {"ranges": absolute_cell, "fields": "sheets/data/rowData/values/note"}
+        response = self.client.request("get", url, params=params)
         response.raise_for_status()
         response_json = response.json()
 
         try:
-            note = response_json['sheets'][0]['data'][0]['rowData'][0]['values'][0]['note']
+            note = response_json["sheets"][0]["data"][0]["rowData"][0]["values"][0][
+                "note"
+            ]
         except (IndexError, KeyError):
-            note = ''
+            note = ""
 
         return note
 
@@ -2127,16 +1606,8 @@ class Worksheet(object):
                             "startColumnIndex": startColumn - 1,
                             "endColumnIndex": startColumn,
                         },
-                        "rows": [
-                            {
-                                "values": [
-                                    {
-                                        "note": content
-                                    }
-                                ]
-                            }
-                        ],
-                        "fields": "note"
+                        "rows": [{"values": [{"note": content}]}],
+                        "fields": "note",
                     }
                 }
             ]
@@ -2182,70 +1653,3 @@ class Worksheet(object):
         """
         # set the note to <empty string> will clear it
         self.update_note(cell, "")
-
-
-class Cell(object):
-    """An instance of this class represents a single cell
-    in a :class:`worksheet <gspread.models.Worksheet>`.
-    """
-
-    def __init__(self, row, col, value=''):
-        self._row = row
-        self._col = col
-
-        #: Value of the cell.
-        self.value = value
-
-    @classmethod
-    def from_address(cls, label, value=''):
-        return cls(*a1_to_rowcol(label), value=value)
-
-    def __repr__(self):
-        return '<%s R%sC%s %s>' % (
-            self.__class__.__name__,
-            self.row,
-            self.col,
-            repr(self.value),
-        )
-
-    @property
-    def row(self):
-        """Row number of the cell."""
-        return self._row
-
-    @property
-    def col(self):
-        """Column number of the cell."""
-        return self._col
-
-    @property
-    def numeric_value(self):
-        numeric_value = numericise(self.value, default_blank=None)
-
-        # if could not convert, return None
-        if type(numeric_value) == int or type(numeric_value) == float:
-            return numeric_value
-        else:
-            return None
-
-    @property
-    def address(self):
-        """Cell address in A1 notation."""
-        return rowcol_to_a1(self.row, self.col)
-
-    @property
-    def input_value(self):
-        """.. deprecated:: 2.0
-
-        This feature is not supported in Sheets API v4.
-        """
-        import warnings
-
-        warnings.warn(
-            "Cell.input_value is deprecated, "
-            "this feature is not supported in Sheets API v4. "
-            "Please use `value_render_option` when you "
-            "Retrieve `Cell` objects (e.g. in `Worksheet.range()` method).",
-            DeprecationWarning,
-            stacklevel=2,
-        )
